``` Update configs with regulated type and enable tracking
Adjusted the configuration files for preprod, preview, prod, and stage
environments. Changed the `:type` field to "fallback - regulated only"
for consistency in preprod and prod. Introduced 'enabletracking' in
preview and prod, and removed it from stage for appropriate tracking
setup. Updated the default environment return value to 'unknown'
when no environment matches in `initializeSiteConfig`.