feat: align last modified date to bottom of blog item
The blog item cards now use flexbox to position the last modified date
at the bottom. This improves the visual consistency and balance of each
item when blog posts have varying amounts of content.

The "last-modified" class was added to the date paragraph to enable
targeting it with the new flex alignment styles.