refactor: Fix slide builder text container bug
The code changes in slide-builder.js fix a bug in the text container of the slide builder. Previously, the slide title and description were not being appended correctly to the text container, resulting in incorrect rendering. This commit corrects the issue by properly appending the slide title and description elements to the text container, ensuring the correct display of slide content.