feat: Refactor extractSeriesInfo function in Blogroll block
This commit refactors the extractSeriesInfo function in the Blogroll block to include the basePath parameter. The basePath is used to determine the base path of the blog post, which is then used to group and sort the posts. This change improves the accuracy of grouping and sorting blog posts in the Blogroll block.