Update slides.html to reference new JavaScript and CSS files for the React Slide Builder
This commit modifies the slides.html file to update the script and link tags, changing the source of the JavaScript and CSS files to slide-builder-main.js and slide-builder-main.css, respectively. This change ensures that the latest resources are utilized for the React Slide Builder application, contributing to improved functionality and performance.