Update bio.html to refine professional description
This commit modifies the caption in the bio section of the bio.html file, enhancing the description of Tom's professional experience by specifying his role as an Architect in addition to being an IT specialist and programmer. This update aims to provide a clearer representation of Tom's expertise and contributions.